+++ TWLight/users/helpers/authorizations.py
@@ -2,7 +2,7 @@ from TWLight.resources.models import Partner
 from TWLight.users.models import Authorization
 
 
-# We have to put this in a new bundle_authorizations.py helper
+# We have to put this in a new authorizations.py helper
 # rather than validation.py to avoid circular imports
 def get_all_bundle_authorizations():
     """
