+++ TWLight/users/signals.py
@@ -3,7 +3,7 @@ from django.conf import settings
 from django.contrib.auth.models import User
 from django.dispatch import receiver, Signal
 from django.db.models.signals import pre_save, post_save, post_delete
-from TWLight.users.helpers.bundle_authorizations import get_all_bundle_authorizations
+from TWLight.users.helpers.authorizations import get_all_bundle_authorizations
 from TWLight.users.models import Authorization, UserProfile
 from TWLight.resources.models import Partner, Stream
