+++ TWLight/users/tests.py
@@ -24,7 +24,7 @@ from TWLight.resources.tests import EditorCraftRoom
 from . import views
 from .oauth import OAuthBackend
 from .helpers.validation import validate_partners
-from .helpers.bundle_authorizations import get_all_bundle_authorizations
+from .helpers.authorizations import get_all_bundle_authorizations
 from .helpers.wiki_list import WIKIS, LANGUAGE_CODES
 from .factories import EditorFactory, UserFactory
 from .groups import get_coordinators, get_restricted
